chore : collisionmatrix Update

+++ Game/Project/Engine/EHAnimation2D.cpp
@@ -101,7 +101,7 @@ void Animation2D::LateUpdate()
 		++m_CurFrame;
 		if (m_FrameInfo.size() <= m_CurFrame)
 		{
-			m_CurFrame = m_FrameInfo.size() - 1;
+			m_CurFrame = int(m_FrameInfo.size() - (size_t)1);
 			m_Finsih = TRUE;
 		}
 		m_PlayTime = 0.f;
